+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Support [ipwho.is](https://ipwho.is/)
+### Changed
 ## [0.4] - 2025-06-09
 ### Added
 - Support [ipapi.co](https://ipapi.co/json/)