+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Support [ipwho.is](https://ipwho.is/)
 - Support [ipquery.io](http://api.ipquery.io/?format=json)
+- Support [wtfismyip.com](https://wtfismyip.com/json)
 ### Changed
 ## [0.4] - 2025-06-09
 ### Added