+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `is_ipv6` function
 - `get_private_ipv6` function
 ### Changed
+- Test system modified
+- `README.md` updated
 ## [0.4] - 2025-06-09
 ### Added
 - Support [ipapi.co](https://ipapi.co/json/)