+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support [ipwho.is](https://ipwho.is/)
 - Support [ipquery.io](http://api.ipquery.io/?format=json)
 - Support [wtfismyip.com](https://wtfismyip.com/json)
+- Support [ident.me](https://ident.me/json) IPv6 API
+- Support [tnedi.me](https://tnedi.me/json) IPv6 API
+- Support [ip.sb](https://api.ip.sb/geoip) IPv6 API
 - `is_ipv6` function
 - `get_private_ipv6` function
 - `get_public_ipv6` function