+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Support [tnedi.me](https://tnedi.me/json) IPv6 API
 - Support [ip.sb](https://api.ip.sb/geoip) IPv6 API
 - Support [ipleak.net](https://ipleak.net/json/) IPv6 API
+- Support [my-ip.io](https://www.my-ip.io/) IPv6 API
 - `is_ipv6` function
 - `get_private_ipv6` function
 - `get_public_ipv6` function